% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_to_mean.R
\name{beta_to_mean}
\alias{beta_to_mean}
\title{Get the mean and sd from estimated betas}
\usage{
beta_to_mean(fit, n_states)
}
\arguments{
\item{fit}{fitted model output}

\item{n_states}{number of fitted states}
}
\description{
Get the mean and sd from estimated betas
}
