% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHMMSSF.R
\name{fitHMMSSF}
\alias{fitHMMSSF}
\title{Fit the HMM-SSF}
\usage{
fitHMMSSF(
  ssf_formula,
  tpm_formula = ~1,
  data,
  par0,
  n_states,
  dist = "gamma",
  optim_opts = list(trace = 0, maxit = 50000)
)
}
\arguments{
\item{ssf_formula}{model formula for ssf}

\item{data}{data with columns for ID, stratum, obs,and covariates in formula}

\item{par0}{list of starting values for parameters (betas, alphas)}

\item{n_states}{how many states in the model}

\item{dist}{distribution of the controls (default = "gamma", also "uniform")}

\item{tmp_formula}{formula for transition probabilities (default = ~1)}
}
\description{
Fit the HMM-SSF
}
