% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_par.R
\name{format_par}
\alias{format_par}
\title{Format parameters back into matrix form}
\usage{
format_par(par, n_states, n_ssf_cov, n_tpm_cov)
}
\arguments{
\item{par}{vector of parameters (ssf_par, tpm_par)}

\item{n_states}{number of states in the HMM}

\item{n_ssf_cov}{number of SSF covariates}

\item{n_tpm_cov}{number of TPM covariates}
}
\description{
Format parameters back into matrix form
}
