% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_backward.R
\name{forward_backward}
\alias{forward_backward}
\title{Forward-backward algorithm
(modified from hmmTMB function 'HMM$forward_backward')}
\usage{
forward_backward(mod)
}
\arguments{
\item{ssf_formula}{SSF formula of the HMM-SSF}

\item{tpm_formula}{TP formula}

\item{data}{data used to fit the model}

\item{fit}{fitted model output from fitHMMSSF}

\item{n_states}{number of states}
}
\value{
List with two elements: \code{log_forward}
(log forward probabilities) and \code{log_backward}
(log backward probabilities)
}
\description{
Forward-backward algorithm
(modified from hmmTMB function 'HMM$forward_backward')
}
