% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_controls.R
\name{get_controls}
\alias{get_controls}
\title{Simulate random locations for model fitting}
\usage{
get_controls(obs, n_controls, distr = "uniform")
}
\arguments{
\item{obs}{Data frame of observations (ID, x, y, time)}

\item{n_controls}{Number of control steps ("controls") per observation}

\item{distr}{Name(s) of distribution(s) used to generate control steps.
Spatially uniform ("uniform") points are the default. Otherwise, the distributions
should be entered as a vector of the step length and turning angle distributions
(in that order). If only one distribution is entered, this is used as the
step length distribution. Current options for step length: "gamma", "exponential".
Current options for turning angle: "vm", "uniform", "wrpcauchy".}
}
\description{
Simulate random locations for model fitting
}
