% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_scale.R
\name{get_shape_scale}
\alias{get_shape_scale}
\title{Get the scale and scale from estimated ssf_par}
\usage{
get_shape_scale(mod)
}
\arguments{
\item{mod}{fitted model output}
}
\description{
Get the scale and scale from estimated ssf_par
}
