% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_step_and_angle.R
\name{get_step_and_angle}
\alias{get_step_and_angle}
\title{Get step lengths and turning angles in a case/control dataframe}
\usage{
get_step_and_angle(data)
}
\arguments{
\item{data}{Data frame with columns 'obs' (1: observation, 0: control),
'x', and 'y' (and possibly others)}
}
\value{
Data frame passed as input with two extra columns: step and angle
}
\description{
Get step lengths and turning angles in a case/control dataframe
}
