% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian_CI.R
\name{hessian_CI}
\alias{hessian_CI}
\title{Calculate CIs for the HMM-SSF parameters via the Hessian matrix}
\usage{
hessian_CI(fit, n_states, CI_range = 0.95, ssf_MM, tpm_MM)
}
\arguments{
\item{fit}{fitted model output from optim}

\item{n_states}{number of fitted states}

\item{CI_range}{range of the CIs (default = 0.95)}

\item{ssf_MM}{model matrix for SSF}

\item{tpm_MM}{model matrix of transition probabilities}
}
\description{
Calculate CIs for the HMM-SSF parameters via the Hessian matrix
}
