% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmSSF.R
\name{hmmSSF}
\alias{hmmSSF}
\title{Fit the HMM-SSF}
\usage{
hmmSSF(
  ssf_formula,
  tpm_formula = ~1,
  n_states,
  data,
  ssf_par0,
  tpm_par0 = NULL,
  optim_opts = list(trace = 0, maxit = 50000),
  method = "Nelder-Mead"
)
}
\arguments{
\item{ssf_formula}{Model formula for ssf}

\item{tpm_formula}{Formula for transition probabilities (default = ~1)}

\item{n_states}{Number of HMM states}

\item{data}{Data frame with named columns for ID, stratum, obs, and
covariates in ssf_formula}

\item{ssf_par0}{Matrix of starting values for SSF parameters, with one row
for each covariate and one column for each HMM state.}

\item{tpm_par0}{Matrix of starting values for transition probability
parameters, with one row for each covariate (including the first row for
intercepts) and one column for each off-diagonal transition probability.}

\item{optim_opts}{Additional options to be passed to optim, as a named list
(e.g., maxit, trace)}

\item{method}{Optimisation method used in optim (by default "Nelder-Mead)
but can also take any other option available in optim}
}
\description{
Fit the HMM-SSF
}
