% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_decoding.R
\name{local_decoding}
\alias{local_decoding}
\title{Local decoding using the forward-backward algorithm
(modified from hmmTMB function 'HMM$state_probs')}
\usage{
local_decoding(mod)
}
\arguments{
\item{mod}{Fitted model object, as returned by \code{\link{hmmSSF}}}
}
\value{
Matrix of state probabilities, with one row for each
observation time, and one column for each state. The (i, j)-th
element is the probability of being in state j at time i.
}
\description{
Local decoding using the forward-backward algorithm
(modified from hmmTMB function 'HMM$state_probs')
}
