% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllk.R
\name{nllk}
\alias{nllk}
\title{Negative log likelihood of the HMM-SSF (estimated with forward algorithm)}
\usage{
nllk(par, ssf_MM, tpm_MM, ID, stratum, n_states, sampling_densities, n_obs)
}
\arguments{
\item{par}{Vector of working paramters (ssf_par, tpm_par)}

\item{ssf_MM}{model matrix based on SSF formula}

\item{tpm_MM}{model matrix based on transition probability formula}

\item{ID}{track ID from data}

\item{n_states}{number of states in the HMM}

\item{sampling_densities}{dens of the controls for given sampling distribution}

\item{n_obs}{number of observed locations in the data}

\item{statum}{stratum ID from data}
}
\description{
Negative log likelihood of the HMM-SSF (estimated with forward algorithm)
}
