% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssf.R
\name{plot_ssf}
\alias{plot_ssf}
\title{Plot RSS}
\usage{
plot_ssf(mod, var)
}
\arguments{
\item{mod}{Fitted model returned by \code{\link{fitHMMSSF}}}

\item{var}{Name of variable to plot SSF against}
}
\description{
Plot RSS
}
