% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tpm.R
\name{predict_tpm}
\alias{predict_tpm}
\title{Predict transition probabilities (with uncertainty)}
\usage{
predict_tpm(tpm_formula, new_data, fit, n_states, return_CI = TRUE)
}
\arguments{
\item{...}{}
}
\value{
List with elements 'mle', 'upper', and 'lower'.
Each element is an array, where each layer is a transition
probability matrix corresponding to a row of 'new_data'.
}
\description{
Predict transition probabilities (with uncertainty)
}
