% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_locs.R
\name{random_locs}
\alias{random_locs}
\title{Simulate random locations for model fitting}
\usage{
random_locs(obs, n_controls, distr = "uniform")
}
\arguments{
\item{obs}{Data frame of observations (ID, x, y, time)}

\item{n_controls}{Number of control steps ("controls") per observation}

\item{distr}{Name(s) of distribution(s) used to generate control steps.}
}
\description{
Simulate random locations for model fitting
}
