% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_dens.R
\name{sampling_dens}
\alias{sampling_dens}
\title{Get the densities for each location given the sampling distribution}
\usage{
sampling_dens(data, dist, dist_par = NULL)
}
\arguments{
\item{data}{data with column for step length}

\item{dist}{distribution of the controls}

\item{dist_par}{a vector of distribution parameters (default = NULL)}
}
\description{
Get the densities for each location given the sampling distribution
}
