% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_samp.R
\name{setup_samp}
\alias{setup_samp}
\title{Setup importance sampling distributions}
\usage{
setup_samp(distr = "uniform", obs)
}
\arguments{
\item{distr}{movement distributions for sampling (see get_controls for details)}

\item{obs}{data frame with observed locations (ID, x, y, time)}
}
\description{
Setup importance sampling distributions
}
