% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_controls.R
\name{sim_controls}
\alias{sim_controls}
\title{Simulate control steps}
\usage{
sim_controls(obs, n_controls, step_dist)
}
\arguments{
\item{obs}{Data frame of observations (ID, x, y, time)}

\item{n_controls}{Number of control steps ("controls") per observation}

\item{step_dist}{distribution from which to generate control locations
(can be "uniform" or "gamma")}
}
\description{
Simulate control steps
}
