% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_controls2.R
\name{sim_controls2}
\alias{sim_controls2}
\title{Simulate control steps}
\usage{
sim_controls2(obs, n_controls, dist = "uniform")
}
\arguments{
\item{obs}{Data frame of observations (ID, x, y, time)}

\item{n_controls}{Number of control steps ("controls") per observation}

\item{dist}{Name(s) of distribution(s) used to generate control steps.}
}
\description{
Simulate control steps
}
