% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sim_raster}
\alias{sim_raster}
\title{Simulate raster}
\usage{
sim_raster(rho = 11, lim = c(-30, 30, -30, 30), res = 1)
}
\arguments{
\item{rho}{Correlation parameter}

\item{lim}{Limits of raster}

\item{res}{Resolution}
}
\description{
Simulate raster
}
