% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_decoding.R
\name{viterbi_decoding}
\alias{viterbi_decoding}
\title{Generate most probable state sequence using the viterbi algorithm
(modified from moveHMM function 'viterbi')}
\usage{
viterbi_decoding(mod)
}
\arguments{
\item{mod}{Fitted model object, as returned by \code{\link{hmmSSF}}}
}
\description{
Generate most probable state sequence using the viterbi algorithm
(modified from moveHMM function 'viterbi')
}
