% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_hmmSSF.R
\name{confint.hmmSSF}
\alias{confint.hmmSSF}
\title{Calculate CIs for the HMM-SSF parameters via the Hessian matrix}
\usage{
\method{confint}{hmmSSF}(mod, range = 0.95)
}
\arguments{
\item{mod}{Fitted model output returned by \code{\link{fitHMMSSF}}}

\item{range}{range of the CIs (default = 0.95)}
}
\description{
Calculate CIs for the HMM-SSF parameters via the Hessian matrix
}
