% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHMMSSF.R
\name{fitHMMSSF}
\alias{fitHMMSSF}
\title{Fit the HMM-SSF}
\usage{
fitHMMSSF(
  ssf_formula,
  tpm_formula = ~1,
  n_states,
  data,
  par0,
  optim_opts = list(trace = 0, maxit = 50000)
)
}
\arguments{
\item{ssf_formula}{model formula for ssf}

\item{n_states}{how many states in the model}

\item{data}{data with columns for ID, stratum, obs,and covariates in formula}

\item{par0}{list of starting values for parameters (ssf_par, tpm_par)}

\item{tmp_formula}{formula for transition probabilities (default = ~1)}
}
\description{
Fit the HMM-SSF
}
