% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_par.R
\name{format_par}
\alias{format_par}
\title{Format parameters back into matrix form}
\usage{
format_par(par, n_states, ssf_cov, tpm_cov)
}
\arguments{
\item{par}{Vector of parameters (ssf_par, tpm_par)}

\item{n_states}{Number of states in the HMM}

\item{ssf_cov}{Names of SSF covariates}

\item{tpm_cov}{Names of TPM covariates}
}
\description{
Format parameters back into matrix form
}
