% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_backward.R
\name{forward_backward}
\alias{forward_backward}
\title{Forward-backward algorithm
(modified from hmmTMB function 'HMM$forward_backward')}
\usage{
forward_backward(mod)
}
\arguments{
\item{mod}{Fitted model object, as returned by \code{\link{hmmSSF}}}
}
\value{
List with two elements: \code{log_forward}
(log forward probabilities) and \code{log_backward}
(log backward probabilities)
}
\description{
Forward-backward algorithm
(modified from hmmTMB function 'HMM$forward_backward')
}
