% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_scale.R
\name{get_shape_scale}
\alias{get_shape_scale}
\title{Get the scale and scale from estimated ssf_par}
\usage{
get_shape_scale(fit, n_states)
}
\arguments{
\item{fit}{fitted model output}

\item{n_states}{number of fitted states}
}
\description{
Get the scale and scale from estimated ssf_par
}
