% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmSSF.R
\name{hmmSSF}
\alias{hmmSSF}
\title{Fit the HMM-SSF}
\usage{
hmmSSF(
  ssf_formula,
  tpm_formula = ~1,
  n_states,
  data,
  ssf_par0,
  tpm_par0 = NULL,
  maxit = 1000
)
}
\arguments{
\item{ssf_formula}{Model formula for ssf}

\item{tpm_formula}{Formula for transition probabilities (default = ~1)}

\item{n_states}{Number of HMM states}

\item{data}{Data frame with named columns for ID, stratum, obs, and
covariates in ssf_formula}

\item{ssf_par0}{Matrix of starting values for SSF parameters, with one row
for each covariate and one column for each HMM state.}

\item{tpm_par0}{Matrix of starting values for transition probability
parameters, with one row for each covariate (including the first row for
intercepts) and one column for each off-diagonal transition probability.}

\item{maxit}{the max number of iterations that the optimiser should run for (default = 1e3)}
}
\description{
Fit the HMM-SSF
}
