% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_decoding.R
\name{local_decoding}
\alias{local_decoding}
\title{Local decoding using the forward-backward algorithm
(modified from hmmTMB function 'HMM$state_probs')}
\usage{
local_decoding(ssf_formula, tpm_formula, data, fit, n_states, dist = "gamma")
}
\arguments{
\item{ssf_formula}{SSF formula of the HMM-SSF}

\item{tpm_formula}{TP formula}

\item{data}{data used to fit the model}

\item{fit}{fitted model output from fitHMMSSF}

\item{n_states}{number of states}

\item{dist}{control sampling distribution (default = "gamma")}
}
\value{
Matrix of state probabilities, with one row for each
observation time, and one column for each state. The (i, j)-th
element is the probability of being in state j at time i.
}
\description{
Local decoding using the forward-backward algorithm
(modified from hmmTMB function 'HMM$state_probs')
}
