% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_step_angle.R
\name{plot_step_angle}
\alias{plot_step_angle}
\title{Plot histograms with estimated state densities}
\usage{
plot_step_angle(mod, pal = NULL, as_list = FALSE)
}
\arguments{
\item{mod}{Fitted model object, as returned by \code{\link{fitHMMSSF}}}

\item{pal}{color palette to use (default = NULL)}

\item{as_list}{whether or not to return plots as a list
(default = FALSE to simply plot)}
}
\description{
Plot histograms with estimated state densities
}
