% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_delta.R
\name{predict_delta}
\alias{predict_delta}
\title{Predict stationary state probabilities (with uncertainty)}
\usage{
predict_delta(mod, new_data, return_CI = TRUE)
}
\arguments{
\item{mod}{Fitted model object, as returned by \code{\link{fitHMMSSF}}}

\item{new_data}{Data frame with covariate values to use for prediction}

\item{return_CI}{Logical (default: TRUE). Should confidence intervals be
computed?}
}
\value{
List with elements 'mle', 'upper', and 'lower'.
Each element is an array, where each layer is a transition
probability matrix corresponding to a row of 'new_data'.
}
\description{
Predict stationary state probabilities (with uncertainty)
}
