% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_decoding.R
\name{viterbi_decoding}
\alias{viterbi_decoding}
\title{Generate most probable state sequence using the viterbi algorithm
(modified from moveHMM function 'viterbi')}
\usage{
viterbi_decoding(ssf_formula, tpm_formula, data, fit, n_states, dist = "gamma")
}
\arguments{
\item{ssf_formula}{SSF formula of the HMM-SSF}

\item{tpm_formula}{TP formula}

\item{data}{data used to fit the model}

\item{fit}{fitted model output from fitHMMSSF}

\item{n_states}{number of states}

\item{dist}{control sampling distribution (default = "gamma")}
}
\description{
Generate most probable state sequence using the viterbi algorithm
(modified from moveHMM function 'viterbi')
}
